tensil/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tensil"
version = "0.1.0"
description = "The markdown of spreadsheets — a plain text tabular data format designed for git."
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://tensil.dev"
Repository = "https://github.com/tensil-dev/tensil"

[tool.setuptools.packages.find]
include = ["tensil*"]
